Delete .account-number only if it exists

+++ dashboard/lib/tasks/seed.rake
@@ -285,7 +285,7 @@ namespace :seed do
   end
 
   task test_accounts: :environment do
-    File.delete('test/ui/.account-number')
+    File.delete('test/ui/.account-number') if File.exist?('test/ui/.account-number')
     (1..50).each do |i|
       email = "email_#{i}@testing.xx"
       User.find_by_email_or_hashed_email(email).try(:destroy)
